feat: pet search and adoption logic

+++ server/Models/Ong.cs
@@ -13,8 +13,8 @@ namespace apifmu.Models
 
         public string Address { get; set; }
 
-        public decimal Lat { get; set; }
+        public double Lat { get; set; }
 
-        public decimal Long { get; set; }
+        public double Long { get; set; }
     }
 }
